+++ s/tools/rando/types.ts
@@ -1,0 +1,3 @@
+
+export type Random = () => number
+