+++ s/tools/rando/u32ify.ts
@@ -1,0 +1,5 @@
+
+export function u32ify(fraction: number) {
+	return Math.floor(fraction * 0x100000000)
+}
+